+++ src/components/LatestArticleDrawer.jsx
@@ -1,34 +1,52 @@
 import React, { useEffect, useMemo } from 'react'
 import { Drawer } from 'antd'
 import { useGenerateTokenMutation, useGetLatestArticleQuery } from '../store/featureApiSlice'
-import { getToken, isValidToken } from '../util'
+import { isValidToken } from '../token/tokenManager'
+import { getToken } from '../token/tokenStore'
 
 function LatestArticleDrawer({ isOpen, toggleDrawer }) {
   const [generateToken] = useGenerateTokenMutation()
-  const { data, refetch } = useGetLatestArticleQuery({refetchOnMountOrArgChange:  false})
+  const { data: article, refetch } = useGetLatestArticleQuery(undefined, { skip: false })
 
-  const article = useMemo(() => data?.data, [data])
+  const articleSummary = useMemo(()=>`${article?.body?.slice(0, 197)}...`,[article])
 
   useEffect(() => {
-    if (!isValidToken(getToken())) {
-      generateToken()
-    } else if(!data){ refetch() } 
+    const getTokenAndLoadArticle = async () => {
+      const isTokenValid = isValidToken(getToken())
+      if (!isTokenValid) {
+        await generateToken()
+        refetch()
+      } else if (!article) {
+        refetch()
+      }
+    }
+    getTokenAndLoadArticle()
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isOpen])
+  }, [isOpen, article])
 
   return (
-    <>
-      <Drawer title="Latest Article" placement="left" onClose={toggleDrawer} open={isOpen}>
-        {article && (
-          <>
-            {' '}
-            <p>{article.title} written by <small>{article.author}</small></p>
-            <p>published by <small>{article.published}</small></p>
-            <p>{article.body}</p>
-          </>
-        )}
-      </Drawer>
-    </>
+    <Drawer title="Latest Article" placement="left" onClose={toggleDrawer} open={isOpen}>
+      {article && (
+        <>
+          <p>
+            {article.title} written by <small>{article.author}</small>
+          </p>
+          <p>
+            published by <small>{article.published}</small>
+          </p>
+          <p>{articleSummary}</p>
+        </>
+      )}
+    </Drawer>
   )
 }
 export default LatestArticleDrawer
+
+// 1. Remove redundant tags from Drawer component done
+// 2. limit number of chars in article body
+// 3. make generate token async. done
+// 4. move article data transformation into rtk done
+// 5. variable naming
+// 6. variable for storing expression results done
+// Anthony Amponsah12:42 AM
+// token management refactor
