+++ src/store/AppProvider.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Provider } from 'react-redux'
-import { store } from '../store'
+import { store } from '.'
 
 function AppProvider({children}) {
   return (
