+++ src/token/tokenManager.js
@@ -29,12 +29,5 @@ export function isValidToken(token) {
   }
 }
 
-export function setToken(token) {
-  sessionStorage.setItem('token', token)
-}
 
-export function getToken() {
-  const token = sessionStorage.getItem('token')
-  return token?? null;
-}
